refactor: 여러 곳에서 쓰이는 상수들 Constants 클래스로 분리
